added season pass enum, reworked other enums
