+++ core/enum/region.py
@@ -1,64 +1,7 @@
 from enum import Enum
 from typing import Self
 
-
-class State(Enum):
-    ALABAMA = "AL"
-    ALASKA = "AK"
-    ARIZONA = "AZ"
-    ARKANSAS = "AR"
-    CALIFORNIA = "CA"
-    COLORADO = "CO"
-    CONNECTICUT = "CT"
-    DELAWARE = "DE"
-    FLORIDA = "FL"
-    GEORGIA = "GA"
-    HAWAII = "HI"
-    IDAHO = "ID"
-    ILLINOIS = "IL"
-    INDIANA = "IN"
-    IOWA = "IA"
-    KANSAS = "KS"
-    KENTUCKY = "KY"
-    LOUISIANA = "LA"
-    MAINE = "ME"
-    MARYLAND = "MD"
-    MASSACHUSETTS = "MA"
-    MICHIGAN = "MI"
-    MINNESOTA = "MN"
-    MISSISSIPPI = "MS"
-    MISSOURI = "MO"
-    MONTANA = "MT"
-    NEBRASKA = "NE"
-    NEVADA = "NV"
-    NEW_HAMPSHIRE = "NH"
-    NEW_JERSEY = "NJ"
-    NEW_MEXICO = "NM"
-    NEW_YORK = "NY"
-    NORTH_CAROLINA = "NC"
-    NORTH_DAKOTA = "ND"
-    OHIO = "OH"
-    OKLAHOMA = "OK"
-    OREGON = "OR"
-    PENNSYLVANIA = "PA"
-    RHODE_ISLAND = "RI"
-    SOUTH_CAROLINA = "SC"
-    SOUTH_DAKOTA = "SD"
-    TENNESSEE = "TN"
-    TEXAS = "TX"
-    UTAH = "UT"
-    VERMONT = "VT"
-    VIRGINIA = "VA"
-    WASHINGTON = "WA"
-    WEST_VIRGINIA = "WV"
-    WISCONSIN = "WI"
-    WYOMING = "WY"
-
-    def from_name(name: str) -> Self:
-        try:
-            return State[name.upper().replace(" ", "_")]
-        except KeyError:
-            raise ValueError(f"Unknown state name: {name}")
+from core.enum.state import State
 
 
 class Region(Enum):
