+++ core/enum/season_pass.py
@@ -1,0 +1,12 @@
+from enum import Enum
+
+
+class Season_Pass(Enum):
+    EPIC = "Epic"
+    IKON = "Ikon"
+    INDY = "Indy"
+    POWDER_ALLIANCE = "Powder Alliance"
+    MOUNTAIN_COLLECTIVE = "Mountain Collective"
+    COOPER = "Cooper"
+    FREEDOM = "Freedom"
+    POWER = "Power"
