+++ test/test_core/test_enum/test_season_pass.py
@@ -1,0 +1,12 @@
+import pytest
+
+from core.enum.season_pass import Season_Pass
+
+
+def test_season_pass():
+    assert Season_Pass("Epic") == Season_Pass.EPIC
+
+    with pytest.raises(Exception) as exc_info:
+        Season_Pass("Invalid")
+
+    assert "ValueError" in str(exc_info)
